refactor(server): extract the interactive auth flow into an authflow package
Move the browser-facing authorization flow (/auth, connector and password
login, session SSO, MFA, approval and RP-initiated logout) out of the server
god-object into a dedicated server/authflow package, decomposed into focused
sub-packages rather than one large handler:

- server/authflow/session: the session cookie, SSO lookup and auth-session CRUD
  (session.Manager).
- server/authflow/mfa: the authenticator chain and the TOTP/WebAuthn endpoints
  (mfa.Manager, mfa.Provider and the TOTP/WebAuthn providers).
- server/authflow/web: shared browser infrastructure — HTML error rendering and
  issuer-relative URL building (web.UI), embedded by the components that render.
- authflow.Handler: HTTP orchestration only; it holds the components and
  delegates. Their internals are unexported, so the Handler cannot reach past
  each component's API.
- server/log: the request-scoped logging attributes carried through the context
  (request ID, client IP), in a neutral low-level package the server, the auth
  flow and the CLI logger all reference.

Two data-oriented abstractions keep the flow honest:

- nextAuthStep centralises the post-login decision (MFA chain -> consent ->
  issue) as behaviour-free data, so the login, session-login and approval
  handlers share one decision instead of each re-deriving it.
- responseTypeHandler issues the authorization response with one self-selecting
  handler per response_type (code/token/id_token).

The Handler takes a narrow Config instead of reaching into the Server and mounts
its own routes. session.Config, mfa.Provider and PKCEConfig are owned by the
packages; the server and cmd reference them directly.
