+++ server/authflow/authorize.go
@@ -1,4 +1,4 @@
-package server
+package authflow
 
 // authorize.go handles the /auth authorization endpoint: parsing the request,
 // selecting a connector, and the browser-facing (HTML/redirect) error surface.
@@ -17,7 +17,7 @@ import (
 )
 
 // grantTypeFromAuthRequest determines the grant type from the authorization request parameters.
-func (s *Server) grantTypeFromAuthRequest(r *http.Request) string {
+func (h *Handler) grantTypeFromAuthRequest(r *http.Request) string {
 	redirectURI := r.Form.Get("redirect_uri")
 	if redirectURI == oauth2.DeviceCallbackURI || strings.HasSuffix(redirectURI, oauth2.DeviceCallbackURI) {
 		return oauth2.GrantTypeDeviceCode
@@ -32,26 +32,26 @@ func (s *Server) grantTypeFromAuthRequest(r *http.Request) string {
 }
 
 // handleAuthorization handles the OAuth2 auth endpoint.
-func (s *Server) handleAuthorization(w http.ResponseWriter, r *http.Request) {
+func (h *Handler) handleAuthorization(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 	// Extract the arguments
 	if err := r.ParseForm(); err != nil {
-		s.logger.ErrorContext(r.Context(), "failed to parse arguments", "err", err)
+		h.logger.ErrorContext(r.Context(), "failed to parse arguments", "err", err)
 
-		s.renderError(r, w, http.StatusBadRequest, ErrMsgInvalidRequest)
+		h.RenderError(r, w, http.StatusBadRequest, ErrMsgInvalidRequest)
 		return
 	}
 
 	connectorID := r.Form.Get("connector_id")
-	allConnectors, err := s.storage.ListConnectors(ctx)
+	allConnectors, err := h.storage.ListConnectors(ctx)
 	if err != nil {
-		s.logger.ErrorContext(r.Context(), "failed to get list of connectors", "err", err)
-		s.renderError(r, w, http.StatusInternalServerError, "Failed to retrieve connector list.")
+		h.logger.ErrorContext(r.Context(), "failed to get list of connectors", "err", err)
+		h.RenderError(r, w, http.StatusInternalServerError, "Failed to retrieve connector list.")
 		return
 	}
 
 	// Determine the grant type from the authorization request to filter connectors.
-	grantType := s.grantTypeFromAuthRequest(r)
+	grantType := h.grantTypeFromAuthRequest(r)
 	connectors := make([]storage.Connector, 0, len(allConnectors))
 	for _, c := range allConnectors {
 		if conns.GrantTypeAllowed(c.GrantTypes, grantType) {
@@ -61,15 +61,15 @@ func (s *Server) handleAuthorization(w http.ResponseWriter, r *http.Request) {
 
 	// Filter connectors based on the client's allowed connectors list.
 	// client_id is required per RFC 6749 §4.1.1.
-	client, authErr := s.getClientWithAuthError(ctx, r.Form.Get("client_id"))
+	client, authErr := h.getClientWithAuthError(ctx, r.Form.Get("client_id"))
 	if authErr != nil {
-		s.renderError(r, w, authErr.Status, authErr.Error())
+		h.RenderError(r, w, authErr.Status, authErr.Error())
 		return
 	}
 	connectors = conns.Filter(connectors, client.AllowedConnectors)
 
 	if len(connectors) == 0 {
-		s.renderError(r, w, http.StatusBadRequest, "No connectors available for this client.")
+		h.RenderError(r, w, http.StatusBadRequest, "No connectors available for this client.")
 		return
 	}
 
@@ -85,32 +85,32 @@ func (s *Server) handleAuthorization(w http.ResponseWriter, r *http.Request) {
 	if connectorID != "" {
 		for _, c := range connectors {
 			if c.ID == connectorID {
-				connURL.Path = s.absPath("/auth", url.PathEscape(c.ID))
+				connURL.Path = h.AbsPath("/auth", url.PathEscape(c.ID))
 				http.Redirect(w, r, connURL.String(), http.StatusFound)
 				return
 			}
 		}
-		s.renderError(r, w, http.StatusBadRequest, "Connector ID does not match a valid Connector")
+		h.RenderError(r, w, http.StatusBadRequest, "Connector ID does not match a valid Connector")
 		return
 	}
 
-	if len(connectors) == 1 && !s.alwaysShowLogin {
-		connURL.Path = s.absPath("/auth", url.PathEscape(connectors[0].ID))
+	if len(connectors) == 1 && !h.alwaysShowLogin {
+		connURL.Path = h.AbsPath("/auth", url.PathEscape(connectors[0].ID))
 		http.Redirect(w, r, connURL.String(), http.StatusFound)
 		return
 	}
 
 	// Skip connector selection if a valid session exists, unless prompt=select_account or alwaysShowLogin.
-	if s.sessionConfig != nil {
-		authReq, _, err := s.parseAuthorizationRequest(r)
+	if h.sessions.Enabled() {
+		authReq, _, err := h.parseAuthorizationRequest(r)
 		if err != nil {
-			s.logger.ErrorContext(r.Context(), "failed to parse authorization request", "err", err)
+			h.logger.ErrorContext(r.Context(), "failed to parse authorization request", "err", err)
 
 			switch authErr := err.(type) {
 			case *redirectedAuthErr:
 				authErr.Handler().ServeHTTP(w, r)
 			case *displayedAuthErr:
-				s.renderError(r, w, authErr.Status, err.Error())
+				h.RenderError(r, w, authErr.Status, err.Error())
 			default:
 				panic("unsupported error type")
 			}
@@ -119,19 +119,19 @@ func (s *Server) handleAuthorization(w http.ResponseWriter, r *http.Request) {
 		prompt, err := oauth2.ParsePrompt(authReq.Prompt)
 		if err != nil {
 			// Server error because authReq was validated before saving it to database.
-			s.redirectWithError(w, r, authReq, oauth2.ServerError, "Invalid authentication request")
+			h.redirectWithError(w, r, authReq, oauth2.ServerError, "Invalid authentication request")
 			return
 		}
 
 		// Invalid prompts will be validated and properly redirected later
-		if !s.alwaysShowLogin && !prompt.SelectAccount() {
-			session := s.getValidSession(ctx, w, r)
+		if !h.alwaysShowLogin && !prompt.SelectAccount() {
+			session := h.sessions.ValidSession(ctx, w, r)
 			if session != nil {
 				for _, c := range connectors {
 					if c.ID != session.ConnectorID {
 						continue
 					}
-					connURL.Path = s.absPath("/auth", url.PathEscape(session.ConnectorID))
+					connURL.Path = h.AbsPath("/auth", url.PathEscape(session.ConnectorID))
 					http.Redirect(w, r, connURL.String(), http.StatusFound)
 					return
 				}
@@ -139,14 +139,14 @@ func (s *Server) handleAuthorization(w http.ResponseWriter, r *http.Request) {
 		}
 		if prompt.None() {
 			// Cannot authenticate silently with prompt=none.
-			s.redirectWithError(w, r, authReq, oauth2.LoginRequired, "id_token_hint does not match authenticated user")
+			h.redirectWithError(w, r, authReq, oauth2.LoginRequired, "id_token_hint does not match authenticated user")
 			return
 		}
 	}
 
 	connectorInfos := make([]templates.ConnectorInfo, 0, len(connectors))
 	for _, conn := range connectors {
-		connURL.Path = s.absPath("/auth", url.PathEscape(conn.ID))
+		connURL.Path = h.AbsPath("/auth", url.PathEscape(conn.ID))
 		connectorInfos = append(connectorInfos, templates.ConnectorInfo{
 			ID:   conn.ID,
 			Name: conn.Name,
@@ -155,29 +155,23 @@ func (s *Server) handleAuthorization(w http.ResponseWriter, r *http.Request) {
 		})
 	}
 
-	if err := s.templates.Login(r, w, connectorInfos); err != nil {
-		s.logger.ErrorContext(r.Context(), "server template error", "err", err)
+	if err := h.templates.Login(r, w, connectorInfos); err != nil {
+		h.logger.ErrorContext(r.Context(), "server template error", "err", err)
 	}
 }
 
 // getClientWithAuthError retrieves a client by ID and returns a displayedAuthErr on failure.
 // Invalid client_id is not treated as a redirect error per RFC 6749 §4.1.2.1.
 // https://datatracker.ietf.org/doc/html/rfc6749#section-4.1.2.1
-func (s *Server) getClientWithAuthError(ctx context.Context, clientID string) (storage.Client, *displayedAuthErr) {
-	client, err := s.storage.GetClient(ctx, clientID)
+func (h *Handler) getClientWithAuthError(ctx context.Context, clientID string) (storage.Client, *displayedAuthErr) {
+	client, err := h.storage.GetClient(ctx, clientID)
 	if err != nil {
 		if err == storage.ErrNotFound {
-			s.logger.ErrorContext(ctx, "invalid client_id provided", "client_id", clientID)
+			h.logger.ErrorContext(ctx, "invalid client_id provided", "client_id", clientID)
 			return storage.Client{}, newDisplayedErr(http.StatusBadRequest, "Invalid client_id provided.")
 		}
-		s.logger.ErrorContext(ctx, "failed to get client", "client_id", clientID, "err", err)
+		h.logger.ErrorContext(ctx, "failed to get client", "client_id", clientID, "err", err)
 		return storage.Client{}, newDisplayedErr(http.StatusInternalServerError, "Database error.")
 	}
 	return client, nil
-}
-
-func (s *Server) renderError(r *http.Request, w http.ResponseWriter, status int, description string) {
-	if err := s.templates.Err(r, w, status, description); err != nil {
-		s.logger.ErrorContext(r.Context(), "server template error", "err", err)
-	}
 }
