+++ server/authflow/logout.go
@@ -1,4 +1,4 @@
-package server
+package authflow
 
 import (
 	"context"
@@ -32,11 +32,11 @@ import (
 // HMAC key and logout parameters. Without a session (e.g. already expired, or
 // id_token_hint without a cookie) upstream logout is skipped — this is acceptable
 // because RP-Initiated Logout treats upstream SLO as best-effort.
-func (s *Server) handleLogout(w http.ResponseWriter, r *http.Request) {
+func (h *Handler) handleLogout(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 
 	if r.Method != http.MethodGet && r.Method != http.MethodPost {
-		s.renderError(r, w, http.StatusMethodNotAllowed, "Method not allowed.")
+		h.RenderError(r, w, http.StatusMethodNotAllowed, "Method not allowed.")
 		return
 	}
 
@@ -49,8 +49,8 @@ func (s *Server) handleLogout(w http.ResponseWriter, r *http.Request) {
 	// This follows the OIDC spec recommendation and prevents CSRF via
 	// cross-site image/link requests (e.g. <img src="/logout">).
 	if idTokenHint == "" && r.Method == http.MethodGet {
-		if err := s.templates.Logout(r, w, "", false, true); err != nil {
-			s.logger.ErrorContext(ctx, "server template error", "err", err)
+		if err := h.templates.Logout(r, w, "", false, true); err != nil {
+			h.logger.ErrorContext(ctx, "server template error", "err", err)
 		}
 		return
 	}
@@ -58,24 +58,24 @@ func (s *Server) handleLogout(w http.ResponseWriter, r *http.Request) {
 	var userID, connectorID, clientID string
 
 	if idTokenHint != "" {
-		idToken, err := s.validateIDTokenHint(ctx, idTokenHint)
+		idToken, err := h.validateIDTokenHint(ctx, idTokenHint)
 		if err != nil {
-			s.logger.ErrorContext(ctx, "logout: invalid id_token_hint", "err", err)
-			s.renderError(r, w, http.StatusBadRequest, "Invalid id_token_hint.")
+			h.logger.ErrorContext(ctx, "logout: invalid id_token_hint", "err", err)
+			h.RenderError(r, w, http.StatusBadRequest, "Invalid id_token_hint.")
 			return
 		}
 
 		sub := new(internal.IDTokenSubject)
 		if err := internal.Unmarshal(idToken.Subject, sub); err != nil {
-			s.logger.ErrorContext(ctx, "logout: failed to unmarshal subject", "err", err)
-			s.renderError(r, w, http.StatusBadRequest, "Invalid id_token_hint subject.")
+			h.logger.ErrorContext(ctx, "logout: failed to unmarshal subject", "err", err)
+			h.RenderError(r, w, http.StatusBadRequest, "Invalid id_token_hint subject.")
 			return
 		}
 
 		userID = sub.UserId
 		connectorID = sub.ConnId
 
-		s.logger.DebugContext(ctx, "logout: parsed id_token_hint",
+		h.logger.DebugContext(ctx, "logout: parsed id_token_hint",
 			"user_id", userID, "connector_id", connectorID)
 
 		// When cross-client (trusted peers) scopes are used, the token may have
@@ -85,8 +85,8 @@ func (s *Server) handleLogout(w http.ResponseWriter, r *http.Request) {
 			AuthorizingParty string `json:"azp"`
 		}
 		if err := idToken.Claims(&claims); err != nil {
-			s.logger.ErrorContext(ctx, "logout: failed to decode id_token_hint claims", "err", err)
-			s.renderError(r, w, http.StatusBadRequest, "Invalid id_token_hint.")
+			h.logger.ErrorContext(ctx, "logout: failed to decode id_token_hint claims", "err", err)
+			h.RenderError(r, w, http.StatusBadRequest, "Invalid id_token_hint.")
 			return
 		}
 
@@ -103,15 +103,13 @@ func (s *Server) handleLogout(w http.ResponseWriter, r *http.Request) {
 	// If no id_token_hint, try to identify the user from the session cookie.
 	// This allows logout without a hint when the user has an active session.
 	if userID == "" && connectorID == "" {
-		if cookie, err := r.Cookie(s.sessionConfig.CookieName); err == nil && cookie.Value != "" {
-			if uid, cid, nonce, err := internal.ParseSessionCookie(cookie.Value, s.sessionConfig.CookieEncryptionKey); err == nil {
-				// Verify the session exists and nonce matches before trusting the cookie.
-				if session, err := s.storage.GetAuthSession(ctx, uid, cid); err == nil && subtle.ConstantTimeCompare([]byte(session.Nonce), []byte(nonce)) == 1 {
-					userID = uid
-					connectorID = cid
-					s.logger.DebugContext(ctx, "logout: identified user from session cookie",
-						"user_id", userID, "connector_id", connectorID)
-				}
+		if uid, cid, nonce, ok := h.sessions.ParseCookie(r); ok {
+			// Verify the session exists and nonce matches before trusting the cookie.
+			if session, err := h.storage.GetAuthSession(ctx, uid, cid); err == nil && subtle.ConstantTimeCompare([]byte(session.Nonce), []byte(nonce)) == 1 {
+				userID = uid
+				connectorID = cid
+				h.logger.DebugContext(ctx, "logout: identified user from session cookie",
+					"user_id", userID, "connector_id", connectorID)
 			}
 		}
 	}
@@ -119,44 +117,44 @@ func (s *Server) handleLogout(w http.ResponseWriter, r *http.Request) {
 	// Validate post_logout_redirect_uri against registered client URIs.
 	if postLogoutRedirectURI != "" {
 		if clientID == "" {
-			s.renderError(r, w, http.StatusBadRequest, "post_logout_redirect_uri requires id_token_hint.")
+			h.RenderError(r, w, http.StatusBadRequest, "post_logout_redirect_uri requires id_token_hint.")
 			return
 		}
 
-		client, err := s.storage.GetClient(ctx, clientID)
+		client, err := h.storage.GetClient(ctx, clientID)
 		if err != nil {
-			s.logger.ErrorContext(ctx, "logout: failed to get client", "client_id", clientID, "err", err)
-			s.renderError(r, w, http.StatusBadRequest, "Invalid client.")
+			h.logger.ErrorContext(ctx, "logout: failed to get client", "client_id", clientID, "err", err)
+			h.RenderError(r, w, http.StatusBadRequest, "Invalid client.")
 			return
 		}
 
 		if !slices.Contains(client.PostLogoutRedirectURIs, postLogoutRedirectURI) {
-			s.logger.WarnContext(ctx, "logout: unregistered post_logout_redirect_uri",
+			h.logger.WarnContext(ctx, "logout: unregistered post_logout_redirect_uri",
 				"uri", postLogoutRedirectURI, "client_id", clientID)
-			s.renderError(r, w, http.StatusBadRequest, "Unregistered post_logout_redirect_uri.")
+			h.RenderError(r, w, http.StatusBadRequest, "Unregistered post_logout_redirect_uri.")
 			return
 		}
 	}
 
 	// Revoke refresh tokens (does not touch the auth session or user identity).
 	if userID != "" && connectorID != "" {
-		s.issuer.Refresh.Revoke(ctx, userID, connectorID)
+		h.issuer.Refresh.Revoke(ctx, userID, connectorID)
 	}
 
 	// Try upstream logout. This requires a live auth session to store the HMAC key
 	// and logout parameters. If the session doesn't exist (expired, no cookie, etc.)
 	// upstream logout is skipped — RP-Initiated Logout treats upstream SLO as best-effort.
-	if redirectURL, ok := s.tryUpstreamLogout(ctx, userID, connectorID, postLogoutRedirectURI, state, clientID); ok {
+	if redirectURL, ok := h.tryUpstreamLogout(ctx, userID, connectorID, postLogoutRedirectURI, state, clientID); ok {
 		http.Redirect(w, r, redirectURL, http.StatusSeeOther)
 		return
 	}
 
 	// No upstream logout — delete session now, clear cookie, show page.
-	s.logger.DebugContext(ctx, "logout: completing",
+	h.logger.DebugContext(ctx, "logout: completing",
 		"user_id", userID, "connector_id", connectorID, "client_id", clientID)
-	loggedOut := s.deleteAuthSession(ctx, userID, connectorID)
-	s.clearSessionCookie(w)
-	s.finishLogout(w, r, postLogoutRedirectURI, state, loggedOut)
+	loggedOut := h.deleteAuthSession(ctx, userID, connectorID)
+	h.sessions.ClearCookie(w)
+	h.finishLogout(w, r, postLogoutRedirectURI, state, loggedOut)
 }
 
 // handleLogoutCallback receives the redirect back from the upstream provider
@@ -165,37 +163,31 @@ func (s *Server) handleLogout(w http.ResponseWriter, r *http.Request) {
 // Identity is resolved from the session cookie (HttpOnly, Secure, SameSite=Lax).
 // The session must still exist in storage with a non-nil LogoutState (set before
 // the upstream redirect). After validation, the session is deleted.
-func (s *Server) handleLogoutCallback(w http.ResponseWriter, r *http.Request) {
+func (h *Handler) handleLogoutCallback(w http.ResponseWriter, r *http.Request) {
 	ctx := r.Context()
 
 	// Resolve identity from the session cookie.
-	cookie, err := r.Cookie(s.sessionConfig.CookieName)
-	if err != nil || cookie.Value == "" {
-		s.renderError(r, w, http.StatusBadRequest, "Missing session cookie.")
-		return
-	}
-
-	userID, connectorID, nonce, err := internal.ParseSessionCookie(cookie.Value, s.sessionConfig.CookieEncryptionKey)
-	if err != nil {
-		s.renderError(r, w, http.StatusBadRequest, "Invalid session cookie.")
+	userID, connectorID, nonce, ok := h.sessions.ParseCookie(r)
+	if !ok {
+		h.RenderError(r, w, http.StatusBadRequest, "Invalid session cookie.")
 		return
 	}
 
 	// Load the session and verify nonce.
-	session, err := s.storage.GetAuthSession(ctx, userID, connectorID)
+	session, err := h.storage.GetAuthSession(ctx, userID, connectorID)
 	if err != nil {
-		s.logger.ErrorContext(ctx, "logout callback: session not found", "err", err)
-		s.renderError(r, w, http.StatusBadRequest, "Session not found.")
+		h.logger.ErrorContext(ctx, "logout callback: session not found", "err", err)
+		h.RenderError(r, w, http.StatusBadRequest, "Session not found.")
 		return
 	}
 
 	if subtle.ConstantTimeCompare([]byte(session.Nonce), []byte(nonce)) != 1 {
-		s.renderError(r, w, http.StatusBadRequest, "Invalid session.")
+		h.RenderError(r, w, http.StatusBadRequest, "Invalid session.")
 		return
 	}
 
 	if session.LogoutState == nil {
-		s.renderError(r, w, http.StatusBadRequest, "No logout in progress.")
+		h.RenderError(r, w, http.StatusBadRequest, "No logout in progress.")
 		return
 	}
 
@@ -203,11 +195,11 @@ func (s *Server) handleLogoutCallback(w http.ResponseWriter, r *http.Request) {
 
 	// Let the connector validate the upstream logout response if it supports it.
 	if ls.ConnectorID != "" {
-		conn, err := s.connectors.Get(ctx, ls.ConnectorID)
+		conn, err := h.connectors.Get(ctx, ls.ConnectorID)
 		if err == nil {
 			if logoutConn, ok := conn.Connector.(connector.LogoutCallbackConnector); ok {
 				if err := logoutConn.HandleLogoutCallback(ctx, r); err != nil {
-					s.logger.ErrorContext(ctx, "logout: upstream logout response validation failed",
+					h.logger.ErrorContext(ctx, "logout: upstream logout response validation failed",
 						"connector_id", ls.ConnectorID, "err", err)
 				}
 			}
@@ -215,14 +207,14 @@ func (s *Server) handleLogoutCallback(w http.ResponseWriter, r *http.Request) {
 	}
 
 	// Session kept alive until now — delete it and clear the cookie.
-	s.deleteAuthSession(ctx, userID, connectorID)
-	s.clearSessionCookie(w)
-	s.finishLogout(w, r, ls.PostLogoutRedirectURI, ls.State, true)
+	h.deleteAuthSession(ctx, userID, connectorID)
+	h.sessions.ClearCookie(w)
+	h.finishLogout(w, r, ls.PostLogoutRedirectURI, ls.State, true)
 }
 
 // finishLogout renders the logout page with a "Back to Application" link.
 // loggedOut indicates whether an active session was actually terminated.
-func (s *Server) finishLogout(w http.ResponseWriter, r *http.Request, postLogoutRedirectURI, state string, loggedOut bool) {
+func (h *Handler) finishLogout(w http.ResponseWriter, r *http.Request, postLogoutRedirectURI, state string, loggedOut bool) {
 	var backURL string
 	if postLogoutRedirectURI != "" {
 		u, err := url.Parse(postLogoutRedirectURI)
@@ -236,8 +228,8 @@ func (s *Server) finishLogout(w http.ResponseWriter, r *http.Request, postLogout
 		}
 	}
 
-	if err := s.templates.Logout(r, w, backURL, loggedOut, false); err != nil {
-		s.logger.ErrorContext(r.Context(), "server template error", "err", err)
+	if err := h.templates.Logout(r, w, backURL, loggedOut, false); err != nil {
+		h.logger.ErrorContext(r.Context(), "server template error", "err", err)
 	}
 }
 
@@ -245,12 +237,12 @@ func (s *Server) finishLogout(w http.ResponseWriter, r *http.Request, postLogout
 // It stores LogoutState in the auth session before redirecting so the callback can
 // read it back. Returns the redirect URL and true on success, or ("", false) if
 // upstream logout is not possible (no session, connector doesn't support it, etc.).
-func (s *Server) tryUpstreamLogout(ctx context.Context, userID, connectorID, postLogoutRedirectURI, state, clientID string) (string, bool) {
+func (h *Handler) tryUpstreamLogout(ctx context.Context, userID, connectorID, postLogoutRedirectURI, state, clientID string) (string, bool) {
 	if connectorID == "" {
 		return "", false
 	}
 
-	conn, err := s.connectors.Get(ctx, connectorID)
+	conn, err := h.connectors.Get(ctx, connectorID)
 	if err != nil {
 		return "", false
 	}
@@ -261,15 +253,15 @@ func (s *Server) tryUpstreamLogout(ctx context.Context, userID, connectorID, pos
 	}
 
 	// Check that the session exists — we need it to store logout state.
-	_, err = s.storage.GetAuthSession(ctx, userID, connectorID)
+	_, err = h.storage.GetAuthSession(ctx, userID, connectorID)
 	if err != nil {
-		s.logger.DebugContext(ctx, "logout: no auth session for upstream logout, skipping",
+		h.logger.DebugContext(ctx, "logout: no auth session for upstream logout, skipping",
 			"user_id", userID, "connector_id", connectorID)
 		return "", false
 	}
 
 	// Store logout parameters in the session.
-	if err := s.storage.UpdateAuthSession(ctx, userID, connectorID, func(old storage.AuthSession) (storage.AuthSession, error) {
+	if err := h.storage.UpdateAuthSession(ctx, userID, connectorID, func(old storage.AuthSession) (storage.AuthSession, error) {
 		old.LogoutState = &storage.LogoutState{
 			PostLogoutRedirectURI: postLogoutRedirectURI,
 			State:                 state,
@@ -278,14 +270,14 @@ func (s *Server) tryUpstreamLogout(ctx context.Context, userID, connectorID, pos
 		}
 		return old, nil
 	}); err != nil {
-		s.logger.ErrorContext(ctx, "logout: failed to save logout state", "err", err)
+		h.logger.ErrorContext(ctx, "logout: failed to save logout state", "err", err)
 		return "", false
 	}
 
-	callbackURI := s.absURL("/logout/callback")
+	callbackURI := h.AbsURL("/logout/callback")
 	upstreamURL, err := logoutConn.LogoutURL(ctx, callbackURI)
 	if err != nil {
-		s.logger.ErrorContext(ctx, "logout: upstream connector error", "err", err)
+		h.logger.ErrorContext(ctx, "logout: upstream connector error", "err", err)
 		return "", false
 	}
 	if upstreamURL == "" {
@@ -294,7 +286,7 @@ func (s *Server) tryUpstreamLogout(ctx context.Context, userID, connectorID, pos
 
 	u, err := url.Parse(upstreamURL)
 	if err != nil {
-		s.logger.ErrorContext(ctx, "logout: failed to parse upstream URL", "err", err)
+		h.logger.ErrorContext(ctx, "logout: failed to parse upstream URL", "err", err)
 		return "", false
 	}
 
@@ -302,16 +294,16 @@ func (s *Server) tryUpstreamLogout(ctx context.Context, userID, connectorID, pos
 }
 
 // deleteAuthSession deletes the session and returns true if it existed.
-func (s *Server) deleteAuthSession(ctx context.Context, userID, connectorID string) bool {
+func (h *Handler) deleteAuthSession(ctx context.Context, userID, connectorID string) bool {
 	if userID == "" || connectorID == "" {
 		return false
 	}
-	if err := s.storage.DeleteAuthSession(ctx, userID, connectorID); err != nil {
+	if err := h.storage.DeleteAuthSession(ctx, userID, connectorID); err != nil {
 		if !errors.Is(err, storage.ErrNotFound) {
-			s.logger.ErrorContext(ctx, "logout: failed to delete auth session", "err", err)
+			h.logger.ErrorContext(ctx, "logout: failed to delete auth session", "err", err)
 		}
 		return false
 	}
-	s.logger.InfoContext(ctx, "logout successful", "user_id", userID, "connector_id", connectorID)
+	h.logger.InfoContext(ctx, "logout successful", "user_id", userID, "connector_id", connectorID)
 	return true
 }
