+++ server/authflow/mfa/mfa.go
@@ -1,4 +1,6 @@
-package server
+// Package mfa owns dex's MFA domain: the authenticator chain and the TOTP
+// and WebAuthn endpoints. The interactive auth flow delegates to a Manager.
+package mfa
 
 import (
 	"bytes"
@@ -8,6 +10,7 @@ import (
 	"errors"
 	"fmt"
 	"image/png"
+	"log/slog"
 	"net/http"
 	"net/url"
 	"time"
@@ -15,7 +18,10 @@ import (
 	"github.com/pquerna/otp"
 	"github.com/pquerna/otp/totp"
 
+	"github.com/dexidp/dex/server/authflow/web"
+	"github.com/dexidp/dex/server/connectors"
 	"github.com/dexidp/dex/server/internal"
+	"github.com/dexidp/dex/server/templates"
 	"github.com/dexidp/dex/storage"
 )
 
@@ -66,8 +72,8 @@ func validateTOTPCode(secret, code string, now time.Time, lastCounter int64) (ok
 	return false, 0
 }
 
-// MFAProvider is a pluggable multi-factor authentication method.
-type MFAProvider interface {
+// Provider is a pluggable multi-factor authentication method.
+type Provider interface {
 	// Type returns the authenticator type identifier (e.g., "TOTP").
 	Type() string
 	// EnabledForConnectorType returns true if this provider applies to the given connector type.
@@ -117,42 +123,70 @@ type mfaRequestContext struct {
 
 // validateMFARequest performs common MFA request validation: HMAC check, auth request
 // lookup, user identity lookup, and approval URL generation.
-func (s *Server) validateMFARequest(w http.ResponseWriter, r *http.Request) (*mfaRequestContext, bool) {
+// Manager owns the MFA domain: the authenticator chain, the TOTP and WebAuthn
+// endpoints, and challenge lifecycle. It embeds web for error rendering and URL
+// building; the Handler delegates the MFA endpoints and chain lookups to it.
+type Manager struct {
+	*web.UI
+	storage         storage.Storage
+	templates       *templates.Templates
+	logger          *slog.Logger
+	mfaProviders    map[string]Provider
+	defaultMFAChain []string
+	now             func() time.Time
+	connectors      *connectors.Cache
+}
+
+// New builds an MFA Manager.
+func New(ui *web.UI, store storage.Storage, tmpls *templates.Templates, logger *slog.Logger, providers map[string]Provider, defaultChain []string, now func() time.Time, conns *connectors.Cache) *Manager {
+	return &Manager{
+		UI:              ui,
+		storage:         store,
+		templates:       tmpls,
+		logger:          logger,
+		mfaProviders:    providers,
+		defaultMFAChain: defaultChain,
+		now:             now,
+		connectors:      conns,
+	}
+}
+
+func (m *Manager) validateMFARequest(w http.ResponseWriter, r *http.Request) (*mfaRequestContext, bool) {
 	macEncoded := r.FormValue("hmac")
 	if macEncoded == "" {
-		s.renderError(r, w, http.StatusUnauthorized, "Unauthorized request.")
+		m.RenderError(r, w, http.StatusUnauthorized, "Unauthorized request.")
 		return nil, false
 	}
 
 	ctx := r.Context()
 
-	authReq, err := s.storage.GetAuthRequest(ctx, r.FormValue("req"))
+	authReq, err := m.storage.GetAuthRequest(ctx, r.FormValue("req"))
 	if err != nil {
-		s.logger.ErrorContext(ctx, "failed to get auth request", "err", err)
-		s.renderError(r, w, http.StatusInternalServerError, "Database error.")
+		m.logger.ErrorContext(ctx, "failed to get auth request", "err", err)
+		m.RenderError(r, w, http.StatusInternalServerError, "Database error.")
 		return nil, false
 	}
 	if !authReq.LoggedIn {
-		s.logger.ErrorContext(ctx, "auth request does not have an identity for MFA verification")
-		s.renderError(r, w, http.StatusInternalServerError, "Login process not yet finalized.")
+		m.logger.ErrorContext(ctx, "auth request does not have an identity for MFA verification")
+		m.RenderError(r, w, http.StatusInternalServerError, "Login process not yet finalized.")
 		return nil, false
 	}
 
 	authenticatorID := r.FormValue("authenticator")
 
 	if !internal.VerifyHMAC(authReq.HMACKey, macEncoded, authReq.ID, authenticatorID) {
-		s.renderError(r, w, http.StatusUnauthorized, "Unauthorized request.")
+		m.RenderError(r, w, http.StatusUnauthorized, "Unauthorized request.")
 		return nil, false
 	}
 
-	identity, err := s.storage.GetUserIdentity(ctx, authReq.Claims.UserID, authReq.ConnectorID)
+	identity, err := m.storage.GetUserIdentity(ctx, authReq.Claims.UserID, authReq.ConnectorID)
 	if err != nil {
-		s.logger.ErrorContext(ctx, "failed to get user identity", "err", err)
-		s.renderError(r, w, http.StatusInternalServerError, "Database error.")
+		m.logger.ErrorContext(ctx, "failed to get user identity", "err", err)
+		m.RenderError(r, w, http.StatusInternalServerError, "Database error.")
 		return nil, false
 	}
 
-	approvalURL := s.buildApprovalURL(authReq)
+	approvalURL := m.BuildApprovalURL(authReq)
 
 	if authReq.MFAValidated {
 		http.Redirect(w, r, approvalURL, http.StatusSeeOther)
@@ -167,33 +201,33 @@ func (s *Server) validateMFARequest(w http.ResponseWriter, r *http.Request) (*mf
 	}, true
 }
 
-func (s *Server) handleTOTP(w http.ResponseWriter, r *http.Request) {
-	mfa, ok := s.validateMFARequest(w, r)
+func (m *Manager) HandleTOTP(w http.ResponseWriter, r *http.Request) {
+	mfa, ok := m.validateMFARequest(w, r)
 	if !ok {
 		return
 	}
 
-	provider, ok := s.mfaProviders[mfa.authenticatorID]
+	provider, ok := m.mfaProviders[mfa.authenticatorID]
 	if !ok {
-		s.renderError(r, w, http.StatusBadRequest, "Unknown authenticator.")
+		m.RenderError(r, w, http.StatusBadRequest, "Unknown authenticator.")
 		return
 	}
 	totpProvider, ok := provider.(*TOTPProvider)
 	if !ok {
-		s.renderError(r, w, http.StatusBadRequest, "Not a TOTP authenticator.")
+		m.RenderError(r, w, http.StatusBadRequest, "Not a TOTP authenticator.")
 		return
 	}
 
-	s.handleTOTPVerify(w, r, r.Context(), mfa.authReq, mfa.identity, mfa.authenticatorID, totpProvider, mfa.approvalURL)
+	m.handleTOTPVerify(w, r, r.Context(), mfa.authReq, mfa.identity, mfa.authenticatorID, totpProvider, mfa.approvalURL)
 }
 
-func (s *Server) handleWebAuthn(w http.ResponseWriter, r *http.Request) {
+func (m *Manager) HandleWebAuthn(w http.ResponseWriter, r *http.Request) {
 	if r.Method != http.MethodGet {
-		s.renderError(r, w, http.StatusMethodNotAllowed, "Unsupported request method.")
+		m.RenderError(r, w, http.StatusMethodNotAllowed, "Unsupported request method.")
 		return
 	}
 
-	mfa, ok := s.validateMFARequest(w, r)
+	mfa, ok := m.validateMFARequest(w, r)
 	if !ok {
 		return
 	}
@@ -206,13 +240,13 @@ func (s *Server) handleWebAuthn(w http.ResponseWriter, r *http.Request) {
 		mode = "register"
 	}
 
-	if err := s.templates.WebAuthnVerify(r, w, mode, mfa.authenticatorID); err != nil {
-		s.logger.ErrorContext(r.Context(), "server template error", "err", err)
+	if err := m.templates.WebAuthnVerify(r, w, mode, mfa.authenticatorID); err != nil {
+		m.logger.ErrorContext(r.Context(), "server template error", "err", err)
 	}
 }
 
 // handleTOTPVerify handles TOTP enrollment and verification.
-func (s *Server) handleTOTPVerify(w http.ResponseWriter, r *http.Request, ctx context.Context,
+func (m *Manager) handleTOTPVerify(w http.ResponseWriter, r *http.Request, ctx context.Context,
 	authReq storage.AuthRequest, identity storage.UserIdentity,
 	authenticatorID string, totpProvider *TOTPProvider, returnURL string,
 ) {
@@ -226,8 +260,8 @@ func (s *Server) handleTOTPVerify(w http.ResponseWriter, r *http.Request, ctx co
 			// enrollment multiple times without completing it, old secrets accumulate.
 			generated, err := totpProvider.generate(authReq.ConnectorID, authReq.Claims.Email)
 			if err != nil {
-				s.logger.ErrorContext(ctx, "failed to generate TOTP key", "err", err)
-				s.renderError(r, w, http.StatusInternalServerError, "Internal server error.")
+				m.logger.ErrorContext(ctx, "failed to generate TOTP key", "err", err)
+				m.RenderError(r, w, http.StatusInternalServerError, "Internal server error.")
 				return
 			}
 
@@ -236,23 +270,23 @@ func (s *Server) handleTOTPVerify(w http.ResponseWriter, r *http.Request, ctx co
 				Type:            "TOTP",
 				Secret:          generated.String(),
 				Confirmed:       false,
-				CreatedAt:       s.now(),
+				CreatedAt:       m.now(),
 			}
 
-			if err := s.storage.UpdateUserIdentity(ctx, authReq.Claims.UserID, authReq.ConnectorID, func(old storage.UserIdentity) (storage.UserIdentity, error) {
+			if err := m.storage.UpdateUserIdentity(ctx, authReq.Claims.UserID, authReq.ConnectorID, func(old storage.UserIdentity) (storage.UserIdentity, error) {
 				if old.MFASecrets == nil {
 					old.MFASecrets = make(map[string]*storage.MFASecret)
 				}
 				old.MFASecrets[authenticatorID] = secret
 				return old, nil
 			}); err != nil {
-				s.logger.ErrorContext(ctx, "failed to store MFA secret", "err", err)
-				s.renderError(r, w, http.StatusInternalServerError, "Internal server error.")
+				m.logger.ErrorContext(ctx, "failed to store MFA secret", "err", err)
+				m.RenderError(r, w, http.StatusInternalServerError, "Internal server error.")
 				return
 			}
 		}
 
-		s.renderTOTPPage(secret, false, totpProvider.issuer, authReq.ConnectorID, w, r)
+		m.renderTOTPPage(secret, false, totpProvider.issuer, authReq.ConnectorID, w, r)
 
 	case http.MethodPost:
 		// TODO(nabokihms): this endpoint should be protected with a rate limit (like the auth endpoint).
@@ -261,21 +295,21 @@ func (s *Server) handleTOTPVerify(w http.ResponseWriter, r *http.Request, ctx co
 		//
 		// For now the best way is to use external rate limiting solutions.
 		if secret == nil || secret.Secret == "" {
-			s.renderError(r, w, http.StatusBadRequest, "MFA not enrolled.")
+			m.RenderError(r, w, http.StatusBadRequest, "MFA not enrolled.")
 			return
 		}
 
 		code := r.FormValue("totp")
 		generated, err := otp.NewKeyFromURL(secret.Secret)
 		if err != nil {
-			s.logger.ErrorContext(ctx, "failed to load TOTP key", "err", err)
-			s.renderError(r, w, http.StatusInternalServerError, "Internal server error.")
+			m.logger.ErrorContext(ctx, "failed to load TOTP key", "err", err)
+			m.RenderError(r, w, http.StatusInternalServerError, "Internal server error.")
 			return
 		}
 
-		ok, counter := validateTOTPCode(generated.Secret(), code, s.now(), secret.LastUsedCounter)
+		ok, counter := validateTOTPCode(generated.Secret(), code, m.now(), secret.LastUsedCounter)
 		if !ok {
-			s.renderTOTPPage(secret, true, totpProvider.issuer, authReq.ConnectorID, w, r)
+			m.renderTOTPPage(secret, true, totpProvider.issuer, authReq.ConnectorID, w, r)
 			return
 		}
 
@@ -283,9 +317,9 @@ func (s *Server) handleTOTPVerify(w http.ResponseWriter, r *http.Request, ctx co
 		// confirm the secret on first successful use. The counter is re-checked
 		// against the latest stored value inside the updater so two concurrent
 		// requests with the same code cannot both succeed. This burn commits
-		// before completeMFAStep marks the challenge passed, so a code can never
+		// before CompleteStep marks the challenge passed, so a code can never
 		// pass the challenge without its counter being recorded first.
-		if err := s.storage.UpdateUserIdentity(ctx, authReq.Claims.UserID, authReq.ConnectorID, func(old storage.UserIdentity) (storage.UserIdentity, error) {
+		if err := m.storage.UpdateUserIdentity(ctx, authReq.Claims.UserID, authReq.ConnectorID, func(old storage.UserIdentity) (storage.UserIdentity, error) {
 			sec := old.MFASecrets[authenticatorID]
 			if sec == nil {
 				return old, errTOTPNotEnrolled
@@ -298,31 +332,31 @@ func (s *Server) handleTOTPVerify(w http.ResponseWriter, r *http.Request, ctx co
 			return old, nil
 		}); err != nil {
 			if errors.Is(err, errTOTPReplay) || errors.Is(err, errTOTPNotEnrolled) {
-				s.renderTOTPPage(secret, true, totpProvider.issuer, authReq.ConnectorID, w, r)
+				m.renderTOTPPage(secret, true, totpProvider.issuer, authReq.ConnectorID, w, r)
 				return
 			}
-			s.logger.ErrorContext(ctx, "failed to update MFA secret", "err", err)
-			s.renderError(r, w, http.StatusInternalServerError, "Internal server error.")
+			m.logger.ErrorContext(ctx, "failed to update MFA secret", "err", err)
+			m.RenderError(r, w, http.StatusInternalServerError, "Internal server error.")
 			return
 		}
 
-		redirectURL, err := s.completeMFAStep(ctx, authReq, authenticatorID)
+		redirectURL, err := m.CompleteStep(ctx, authReq, authenticatorID)
 		if err != nil {
-			s.logger.ErrorContext(ctx, "failed to complete MFA step", "err", err)
-			s.renderError(r, w, http.StatusInternalServerError, "Internal server error.")
+			m.logger.ErrorContext(ctx, "failed to complete MFA step", "err", err)
+			m.RenderError(r, w, http.StatusInternalServerError, "Internal server error.")
 			return
 		}
 
-		// completeMFAStep returns either the next MFA step URL or the approval URL.
+		// CompleteStep returns either the next MFA step URL or the approval URL.
 		// Redirect in both cases — the approval handler handles skipApproval logic.
 		http.Redirect(w, r, redirectURL, http.StatusSeeOther)
 
 	default:
-		s.renderError(r, w, http.StatusBadRequest, "Unsupported request method.")
+		m.RenderError(r, w, http.StatusBadRequest, "Unsupported request method.")
 	}
 }
 
-func (s *Server) renderTOTPPage(secret *storage.MFASecret, lastFail bool, issuer, connectorID string, w http.ResponseWriter, r *http.Request) {
+func (m *Manager) renderTOTPPage(secret *storage.MFASecret, lastFail bool, issuer, connectorID string, w http.ResponseWriter, r *http.Request) {
 	// Prevent browser from caching the TOTP page (contains QR code with secret).
 	w.Header().Set("Cache-Control", "no-store")
 	var qrCode string
@@ -330,13 +364,13 @@ func (s *Server) renderTOTPPage(secret *storage.MFASecret, lastFail bool, issuer
 		var err error
 		qrCode, err = generateTOTPQRCode(secret.Secret)
 		if err != nil {
-			s.logger.ErrorContext(r.Context(), "failed to generate QR code", "err", err)
-			s.renderError(r, w, http.StatusInternalServerError, "Internal server error.")
+			m.logger.ErrorContext(r.Context(), "failed to generate QR code", "err", err)
+			m.RenderError(r, w, http.StatusInternalServerError, "Internal server error.")
 			return
 		}
 	}
-	if err := s.templates.TOTPVerify(r, w, r.URL.String(), issuer, connectorID, qrCode, lastFail); err != nil {
-		s.logger.ErrorContext(r.Context(), "server template error", "err", err)
+	if err := m.templates.TOTPVerify(r, w, r.URL.String(), issuer, connectorID, qrCode, lastFail); err != nil {
+		m.logger.ErrorContext(r.Context(), "server template error", "err", err)
 	}
 }
 
@@ -359,15 +393,15 @@ func generateTOTPQRCode(keyURL string) (string, error) {
 	return base64.StdEncoding.EncodeToString(buf.Bytes()), nil
 }
 
-// mfaChainForClient returns the MFA chain for a client filtered by connector type,
+// ChainForClient returns the MFA chain for a client filtered by connector type,
 // falling back to the server's defaultMFAChain if the client has none.
 // Returns nil if no MFA is configured/applicable.
-func (s *Server) mfaChainForClient(ctx context.Context, clientID, connectorID string) ([]string, error) {
-	if len(s.mfaProviders) == 0 {
+func (m *Manager) ChainForClient(ctx context.Context, clientID, connectorID string) ([]string, error) {
+	if len(m.mfaProviders) == 0 {
 		return nil, nil
 	}
 
-	client, err := s.storage.GetClient(ctx, clientID)
+	client, err := m.storage.GetClient(ctx, clientID)
 	if err != nil {
 		return nil, err
 	}
@@ -376,18 +410,18 @@ func (s *Server) mfaChainForClient(ctx context.Context, clientID, connectorID st
 	// Explicit empty slice ([]string{}) means "no MFA" — don't fall back.
 	source := client.MFAChain
 	if source == nil {
-		source = s.defaultMFAChain
+		source = m.defaultMFAChain
 	}
 
 	// Resolve connector type from connector ID.
-	connectorType, err := s.getConnectorType(ctx, connectorID)
+	connectorType, err := m.getConnectorType(ctx, connectorID)
 	if err != nil {
 		return nil, err
 	}
 
 	var chain []string
 	for _, authID := range source {
-		provider, ok := s.mfaProviders[authID]
+		provider, ok := m.mfaProviders[authID]
 		if ok && provider.EnabledForConnectorType(connectorType) {
 			chain = append(chain, authID)
 		}
@@ -396,8 +430,8 @@ func (s *Server) mfaChainForClient(ctx context.Context, clientID, connectorID st
 }
 
 // getConnectorType returns the type of the connector with the given ID.
-func (s *Server) getConnectorType(ctx context.Context, connectorID string) (string, error) {
-	conn, err := s.connectors.Get(ctx, connectorID)
+func (m *Manager) getConnectorType(ctx context.Context, connectorID string) (string, error) {
+	conn, err := m.connectors.Get(ctx, connectorID)
 	if err != nil {
 		return "", fmt.Errorf("get connector %q: %w", connectorID, err)
 	}
@@ -405,18 +439,18 @@ func (s *Server) getConnectorType(ctx context.Context, connectorID string) (stri
 }
 
 // mfaPagePath returns the page URL path for the given MFA provider type.
-func (s *Server) mfaPagePath(authenticatorID string) string {
-	provider, ok := s.mfaProviders[authenticatorID]
+func (m *Manager) mfaPagePath(authenticatorID string) string {
+	provider, ok := m.mfaProviders[authenticatorID]
 	if ok && provider.Type() == "WebAuthn" {
 		return "/mfa/webauthn"
 	}
 	return "/mfa/totp"
 }
 
-// completeMFAStep checks for the next authenticator in the MFA chain and either
+// CompleteStep checks for the next authenticator in the MFA chain and either
 // returns the URL for the next step or marks MFA as validated and returns the approval URL.
-func (s *Server) completeMFAStep(ctx context.Context, authReq storage.AuthRequest, authenticatorID string) (string, error) {
-	mfaChain, err := s.mfaChainForClient(ctx, authReq.ClientID, authReq.ConnectorID)
+func (m *Manager) CompleteStep(ctx context.Context, authReq storage.AuthRequest, authenticatorID string) (string, error) {
+	mfaChain, err := m.ChainForClient(ctx, authReq.ClientID, authReq.ConnectorID)
 	if err != nil {
 		return "", fmt.Errorf("get MFA chain: %w", err)
 	}
@@ -431,25 +465,25 @@ func (s *Server) completeMFAStep(ctx context.Context, authReq storage.AuthReques
 	}
 
 	if nextAuthenticator != "" {
-		return s.buildMFARedirectURL(authReq, nextAuthenticator), nil
+		return m.BuildRedirectURL(authReq, nextAuthenticator), nil
 	}
 
 	// All authenticators completed — mark as validated.
-	if err := s.storage.UpdateAuthRequest(ctx, authReq.ID, func(old storage.AuthRequest) (storage.AuthRequest, error) {
+	if err := m.storage.UpdateAuthRequest(ctx, authReq.ID, func(old storage.AuthRequest) (storage.AuthRequest, error) {
 		old.MFAValidated = true
 		return old, nil
 	}); err != nil {
 		return "", fmt.Errorf("update auth request: %w", err)
 	}
 
-	return s.buildApprovalURL(authReq), nil
+	return m.BuildApprovalURL(authReq), nil
 }
 
-// buildMFARedirectURL builds an HMAC-protected redirect URL for the given authenticator.
-func (s *Server) buildMFARedirectURL(authReq storage.AuthRequest, authenticatorID string) string {
+// BuildRedirectURL builds an HMAC-protected redirect URL for the given authenticator.
+func (m *Manager) BuildRedirectURL(authReq storage.AuthRequest, authenticatorID string) string {
 	v := url.Values{}
 	v.Set("req", authReq.ID)
 	v.Set("hmac", internal.ComputeHMAC(authReq.HMACKey, authReq.ID, authenticatorID))
 	v.Set("authenticator", authenticatorID)
-	return s.absPath(s.mfaPagePath(authenticatorID)) + "?" + v.Encode()
+	return m.AbsPath(m.mfaPagePath(authenticatorID)) + "?" + v.Encode()
 }
